Convert project to TypeScript
This commit converts the entire project from JavaScript to TypeScript.

The following changes were made:
- Installed TypeScript and necessary type definitions (@types/node, @types/express, @types/morgan, @types/swagger-ui-express, @types/swagger-jsdoc).
- Created and configured a tsconfig.json file.
- Renamed all .js files in the src directory to .ts.
- Updated import paths in all TypeScript files to remove .js extensions and updated Swagger configuration to look for .ts files.
- Updated package.json:
    - Changed the `main` field to `dist/index.js`.
    - Updated the `start` script to `npm run build && node dist/index.js`.
    - Added a `build` script: `tsc`.
- Fixed all type errors by:
    - Adding explicit types for variables, function parameters, and return values.
    - Creating interfaces and types for complex objects (e.g., `ProcessedGroup`).
    - Adjusting import statements for default exports (e.g., `makeWASocket`).
    - Ensuring proper handling of potentially undefined properties (e.g., `lastDisconnect?.error`).
    - Refactoring controller functions to ensure their signatures are compatible with Express route handlers, resolving TS2769 errors.

The project now successfully compiles with `npx tsc`.
