+++ package.json
@@ -3,10 +3,11 @@
   "version": "1.0.0",
   "type": "module",
   "description": "An REST API using Node & Express. This API is only used to send message to Whatsapp using Baileys socket.",
-  "main": "index.js",
+  "main": "dist/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node src/index.js"
+    "start": "npm run build && node dist/index.js",
+    "build": "tsc"
   },
   "keywords": [],
   "author": "",
@@ -21,5 +22,13 @@
     "qrcode-terminal": "^0.12.0",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1"
+  },
+  "devDependencies": {
+    "@types/express": "^5.0.3",
+    "@types/morgan": "^1.9.10",
+    "@types/node": "^22.15.30",
+    "@types/swagger-jsdoc": "^6.0.4",
+    "@types/swagger-ui-express": "^4.1.8",
+    "typescript": "^5.8.3"
   }
 }
