+++ src/index.ts
@@ -1,4 +1,4 @@
-import app from "./app.js";
+import app from "./app";
 
 const PORT = process.env.PORT || 3000;
